Revert "Refactor chip group into dedicated module"
