+++ public/js/__tests__/createChipGroup.test.js
@@ -1,8 +1,12 @@
+jest.mock('../gcs.js', () => ({ initGcs: jest.fn() }));
+
 describe('createChipGroup', () => {
   let createChipGroup;
 
   beforeAll(() => {
-    ({ createChipGroup } = require('../chipGroup.js'));
+    Object.defineProperty(document, 'readyState', { configurable: true, value: 'loading' });
+    document.body.innerHTML = '<div id="fastGrid"></div><div id="teamGrid"></div>';
+    ({ createChipGroup } = require('../app.js'));
   });
 
   beforeEach(() => {
